requirements.txt
aiohttp==3.6.0
discord.py
dblpy==0.3.3
requests
timeago==1.0.10
tabulate==0.8.3
psutil==5.6.7